+++ vite.config.ts
@@ -6,8 +6,8 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['vitest/**/*.test.ts', 'vitest/**/*.spec.ts'],
-    exclude: ['dist', 'node_modules', 'frontend', 'frontend/**'],
+    include: ['**/vitest/**/*.test.ts', '**/vitest/**/*.spec.ts'],
+    exclude: ['**/dist/', '**/node_modules/', '**/apps/'],
     globals: true,
     clearMocks: true,
     restoreMocks: true,
@@ -16,8 +16,8 @@ export default defineConfig({
     coverage: {
       provider: 'v8', // default, but explicit
       reporter: ['text', 'lcov'],
-      include: ['src/matterbridge.ts'],
-      exclude: ['dist', 'node_modules', 'frontend', 'frontend/**'],
+      include: ['**/vitest/**/*.test.ts', '**/vitest/**/*.spec.ts'],
+      exclude: ['**/dist/', '**/node_modules/', '**/apps/'],
       thresholds: {
         statements: 100,
         branches: 100,